+++ h2gis-functions/src/main/java/org/h2gis/functions/spatial/clusters/ST_ClusterWithin.java
@@ -2,18 +2,18 @@
  * H2GIS is a library that brings spatial support to the H2 Database Engine
  * <a href="http://www.h2database.com">http://www.h2database.com</a>. H2GIS is developed by CNRS
  * <a href="http://www.cnrs.fr/">http://www.cnrs.fr/</a>.
- *
+ * <p>
  * This code is part of the H2GIS project. H2GIS is free software;
  * you can redistribute it and/or modify it under the terms of the GNU
  * Lesser General Public License as published by the Free Software Foundation;
  * version 3.0 of the License.
- *
+ * <p>
  * H2GIS is distributed in the hope that it will be useful, but
  * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
  * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License
  * for more details <http://www.gnu.org/licenses/>.
- *
- *
+ * <p>
+ * <p>
  * For more information, please consult: <a href="http://www.h2gis.org/">http://www.h2gis.org/</a>
  * or contact directly: info_at_h2gis.org
  */
@@ -54,9 +54,9 @@ public class ST_ClusterWithin extends AbstractFunction implements ScalarFunction
      * @return ResultSet with clustered geometries.
      */
     public static ResultSet execute(Connection connection, String tableName, String geomColumn,
-                                    String idColumn,Double eps) throws SQLException {
+                                    String idColumn, Double eps) throws SQLException {
         ClusterWithin cluster = new ClusterWithin(
-                connection, tableName, geomColumn, idColumn,eps);
+                connection, tableName, geomColumn, idColumn, eps);
         return cluster.getResultSet();
     }
 }